refactor: Removed savePercentage that was no longer used
